+++ general/CMakeLists.txt
@@ -43,10 +43,8 @@ list(APPEND HDRS
   mm.hpp
   occa.hpp
   okina.hpp
-  openmp.hpp
   optparser.hpp
   osockstream.hpp
-  raja.hpp
   sets.hpp
   socketstream.hpp
   sort_pairs.hpp
