+++ general/cuda.hpp
@@ -21,37 +21,22 @@
 namespace mfem
 {
 
-// *****************************************************************************
 #ifdef MFEM_USE_CUDA
+
 inline void CuCheck(const unsigned int c)
 {
    MFEM_ASSERT(c == cudaSuccess, cudaGetErrorString(cudaGetLastError()));
 }
-template <typename BODY> __global__ static
-void CuKernel(const int N, BODY body)
-{
-   const int k = blockDim.x*blockIdx.x + threadIdx.x;
-   if (k >= N) { return; }
-   body(k);
-}
-template <int BLOCKS, typename DBODY>
-void CuWrap(const int N, DBODY &&d_body)
-{
-   if (N==0) { return; }
-   const int GRID = (N+BLOCKS-1)/BLOCKS;
-   CuKernel<<<GRID,BLOCKS>>>(N,d_body);
-   const cudaError_t last = cudaGetLastError();
-   MFEM_ASSERT(last == cudaSuccess, cudaGetErrorString(last));
-}
-#else // ***********************************************************************
+
+#else
+
 #define __host__
 #define __device__
 #define __constant__
 typedef int CUdevice;
 typedef int CUcontext;
 typedef void* CUstream;
-template <int BLOCKS, typename DBODY>
-void CuWrap(const int N, DBODY &&d_body) {}
+
 #endif // MFEM_USE_CUDA
 
 #if defined(__CUDA_ARCH__)
@@ -70,45 +55,30 @@ template<typename T> inline T AtomicAdd(T* address, T val)
    return *address;
 }
 #endif //__CUDA_ARCH__
-// *****************************************************************************
-// * Allocates device memory
-// *****************************************************************************
+
+/// Allocates device memory
 void* CuMemAlloc(void **d_ptr, size_t bytes);
 
-// *****************************************************************************
-// * Frees device memory
-// *****************************************************************************
+/// Frees device memory
 void* CuMemFree(void *d_ptr);
 
-// *****************************************************************************
-// * Copies memory from Host to Device
-// *****************************************************************************
+/// Copies memory from Host to Device
 void* CuMemcpyHtoD(void *d_dst, const void *h_src, size_t bytes);
 
-// *****************************************************************************
-// * Copies memory from Host to Device
-// *****************************************************************************
+/// Copies memory from Host to Device
 void* CuMemcpyHtoDAsync(void *d_dst, const void *h_src,
                         size_t bytes, void *stream);
 
-// *****************************************************************************
-// * Copies memory from Device to Device
-// *****************************************************************************
+/// Copies memory from Device to Device
 void* CuMemcpyDtoD(void *d_dst, void *d_src, size_t bytes);
 
-// *****************************************************************************
-// * Copies memory from Device to Device
-// *****************************************************************************
+/// Copies memory from Device to Device
 void* CuMemcpyDtoDAsync(void *d_dst, void *d_src, size_t bytes, void *stream);
 
-// *****************************************************************************
-// * Copies memory from Device to Host
-// *****************************************************************************
+/// Copies memory from Device to Host
 void* CuMemcpyDtoH(void *h_dst, void *d_src, size_t bytes);
 
-// *****************************************************************************
-// * Copies memory from Device to Host
-// *****************************************************************************
+/// Copies memory from Device to Host
 void* CuMemcpyDtoHAsync(void *h_dst, void *d_src, size_t bytes, void *stream);
 
 } // namespace mfem
