--- general/openmp.hpp
@@ -1,37 +1,0 @@
-// Copyright (c) 2010, Lawrence Livermore National Security, LLC. Produced at
-// the Lawrence Livermore National Laboratory. LLNL-CODE-443211. All Rights
-// reserved. See file COPYRIGHT for details.
-//
-// This file is part of the MFEM library. For more information and source code
-// availability see http://mfem.org.
-//
-// MFEM is free software; you can redistribute it and/or modify it under the
-// terms of the GNU Lesser General Public License (as published by the Free
-// Software Foundation) version 2.1 dated February 1999.
-
-#ifndef MFEM_OPENMP_HPP
-#define MFEM_OPENMP_HPP
-
-namespace mfem
-{
-
-// *****************************************************************************
-// * Standard OpenMP wrapper
-// *****************************************************************************
-template <typename HBODY>
-void OmpWrap(const int N, HBODY &&h_body)
-{
-#if defined(_OPENMP)
-   #pragma omp parallel for
-   for (int k=0; k<N; k+=1)
-   {
-      h_body(k);
-   }
-#else
-   MFEM_ABORT("OpenMP requested for MFEM but OpenMP is not enabled!");
-#endif
-}
-
-} // namespace mfem
-
-#endif // MFEM_OPENMP_HPP
