--- general/raja.hpp
@@ -1,63 +1,0 @@
-// Copyright (c) 2010, Lawrence Livermore National Security, LLC. Produced at
-// the Lawrence Livermore National Laboratory. LLNL-CODE-443211. All Rights
-// reserved. See file COPYRIGHT for details.
-//
-// This file is part of the MFEM library. For more information and source code
-// availability see http://mfem.org.
-//
-// MFEM is free software; you can redistribute it and/or modify it under the
-// terms of the GNU Lesser General Public License (as published by the Free
-// Software Foundation) version 2.1 dated February 1999.
-
-#ifndef MFEM_RAJA_HPP
-#define MFEM_RAJA_HPP
-
-#ifdef MFEM_USE_RAJA
-#include "RAJA/RAJA.hpp"
-#endif
-
-namespace mfem
-{
-
-// *****************************************************************************
-// * RAJA Cuda wrapper
-// *****************************************************************************
-template <int BLOCKS, typename DBODY>
-void RajaCudaWrap(const int N, DBODY &&d_body)
-{
-#if defined(MFEM_USE_RAJA) && defined(RAJA_ENABLE_CUDA)
-   RAJA::forall<RAJA::cuda_exec<BLOCKS>>(RAJA::RangeSegment(0,N),d_body);
-#else
-   MFEM_ABORT("RAJA::Cuda requested but RAJA::Cuda is not enabled!");
-#endif
-}
-
-// *****************************************************************************
-// * RAJA OpenMP wrapper
-// *****************************************************************************
-template <typename HBODY>
-void RajaOmpWrap(const int N, HBODY &&h_body)
-{
-#if defined(MFEM_USE_RAJA) && defined(RAJA_ENABLE_OPENMP)
-   RAJA::forall<RAJA::omp_parallel_for_exec>(RAJA::RangeSegment(0,N), h_body);
-#else
-   MFEM_ABORT("RAJA::OpenMP requested but RAJA::OpenMP is not enabled!");
-#endif
-}
-
-// *****************************************************************************
-// * RAJA sequential loop wrapper
-// *****************************************************************************
-template <typename HBODY>
-void RajaSeqWrap(const int N, HBODY &&h_body)
-{
-#ifdef MFEM_USE_RAJA
-   RAJA::forall<RAJA::loop_exec>(RAJA::RangeSegment(0,N), h_body);
-#else
-   MFEM_ABORT("RAJA requested but RAJA is not enabled!");
-#endif
-}
-
-} // namespace mfem
-
-#endif // MFEM_RAJA_HPP
